1586 Improve code coverage from 11% to 15%
